vmm: Move GIC unit test to `hypervisor` crate
